Still trying to fix a problem with the DPI

+++ src/main/java/uk/ac/gla/atanaspam/network/utils/ConfiguratorStateKeeper.java
@@ -55,7 +55,7 @@ public class ConfiguratorStateKeeper implements Serializable {
 
     private ArrayList<HashMap<Object, HitCountPair[]>> data;
     private int numOfBolts;
-    private long numberOfPacketsDropped;
+    private double numberOfPacketsDropped;
 
     public ConfiguratorStateKeeper(int numOfBolts){
         numberOfPacketsDropped = 0;
@@ -168,7 +168,7 @@ public class ConfiguratorStateKeeper implements Serializable {
         this.numberOfPacketsDropped++;
     }
 
-    public long getNumberOfPacketsDropped() {
+    public double getNumberOfPacketsDropped() {
         return numberOfPacketsDropped;
     }
